refactor: db in other place
